example/requirements.txt
gensim==3.5.0
tqdm==4.23.4
grpcio==1.13.0
